player movement, block movement, and key retrieval now work correctly
added function to determine validity of a move

movable blocks now destroy blocks they are supposed to, properly move over keys, and do not move when the move is not valid.

keys are now counted when they are generated on load. Each time a key is picked up the key count is reduced. When all the keys are picked up a placerholder event occurs